% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCMapperFunctions.R
\name{CreateClusterHeatmap}
\alias{CreateClusterHeatmap}
\title{CreateClusterHeatmap}
\usage{
CreateClusterHeatmap(images, nb.cells, cluster.map, cluster.ids)
}
\arguments{
\item{images}{list of images in data format}

\item{nb.cells}{list with number of cells per image}

\item{cluster.map}{colors}

\item{cluster.ids}{clusters}
}
\value{
list of images in data format where cell clusters have been replaced by colors
}
\description{
Create a heatmap based on the clusters
}
\keyword{clusters}
\keyword{color}
