% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCMapperFunctions.R
\name{CreateHeatmap}
\alias{CreateHeatmap}
\title{CreateHeatmap}
\usage{
CreateHeatmap(images, nb.cells, counts.ids)
}
\arguments{
\item{images}{list of images in data format}

\item{nb.cells}{list with number of cells per image}

\item{counts.ids}{data table with number of counts for all cells}
}
\value{
list of images in data format where number of counts is indicated
}
\description{
Create a heatmap based on the counts
}
\keyword{cell}
\keyword{counts}
