% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCMapperFunctions.R
\name{DisplayCellsAndCounts}
\alias{DisplayCellsAndCounts}
\title{DisplayCellsAndCounts}
\usage{
DisplayCellsAndCounts(value.image, cell.sel.1, cell.sel.2, cell.sel.3, im.list,
  cb, bcg, ...)
}
\arguments{
\item{value.image}{channel heatmap to display}

\item{cell.sel.1}{list of images in data format containing the 1st cell population to display}

\item{cell.sel.2}{list of images in data format containing the 2nd cell population to display}

\item{cell.sel.3}{list of images in data format containing the 3rd cell population to display}

\item{im.list}{vector with names of images to be displayed}

\item{cb}{list of images in data format with cell borders indicated}

\item{bcg}{brightness / contrast / gamma (e.g. c(0,4,2))}

\item{...}{additional parameters for the label and legend}
}
\value{
nothing is explicitly returned
}
\description{
Display a channel heatmap or an IMC image (white) together with the borders of one to three cell populations (RGB)
}
\keyword{cell}
\keyword{channel}
\keyword{count}
\keyword{display}
\keyword{population}
