% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCMapperFunctions.R
\name{DisplayColorImages}
\alias{DisplayColorImages}
\title{DisplayColorImages}
\usage{
DisplayColorImages(cluster.image, images, borders, bcg = c(0, 1, 1), ...)
}
\arguments{
\item{cluster.image}{list of images in data format where cell ids have been replaced by cluster-dependent colors}

\item{images}{list of images in data format}

\item{borders}{TRUE/FALSE should cell borders be displayed?}

\item{bcg}{brightness / contrast / gamma (default = c(0,1,1))}

\item{...}{additional parameters for the label and legend}
}
\value{
nothing is explicitly returned
}
\description{
Display a color image when custom colors have already been defined
}
\keyword{color}
\keyword{counts}
\keyword{display}
\keyword{image}
