% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCMapperFunctions.R
\name{DisplayIMCandCells}
\alias{DisplayIMCandCells}
\title{DisplayIMCandCells}
\usage{
DisplayIMCandCells(value.image.1, value.image.2, value.image.3, cell.sel.1,
  im.list, cb, bcg1, bcg2, bcg3, ...)
}
\arguments{
\item{value.image.1}{heatmaps for the 1st channel to display}

\item{value.image.2}{heatmaps for the 2nd channel to display}

\item{value.image.3}{heatmaps for the 3rd channel to display}

\item{cell.sel.1}{list of images in data format containing the cell population to display}

\item{im.list}{vector with names of images to be displayed}

\item{cb}{list of images in data format with cell borders indicated}

\item{bcg1}{brightness / contrast / gamma for the 1st image (e.g. = c(0,4,2))}

\item{bcg2}{brightness / contrast / gamma for the 2nd image (e.g. = c(0,4,2))}

\item{bcg3}{brightness / contrast / gamma for the 3rd image (e.g. = c(0,4,2))}

\item{...}{additional parameters for the label and legend}
}
\value{
nothing is explicitly returned
}
\description{
Display one to three IMC images (RGB) together with the borders of one cell population (white)
}
\keyword{IMC}
\keyword{borders}
\keyword{display}
\keyword{image}
