% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCMapperFunctions.R
\name{DisplayorSave}
\alias{DisplayorSave}
\title{DisplayorSave}
\usage{
DisplayorSave(..., image, method = "raster", save = FALSE, path = getwd(),
  filename = "image", type = "png", fun = "", label = TRUE,
  label.txt = NULL, label.color = "red", label.x = 50, label.y = 50,
  label.cex = 3)
}
\arguments{
\item{...}{additional parameters for the legend (see ?legend)}

\item{image}{image to be displayed}

\item{method}{'raster' or 'browser'}

\item{save}{TRUE/FALSE should the image be saved?}

\item{path}{path to the folder where images should be saved}

\item{filename}{optional filename for the images (note: should be modified and used as a suffix or prefix)}

\item{type}{format in which the image should be saved ("png", "tiff" or "jpeg")}

\item{fun}{name of the parent function}

\item{label}{TRUE/FALSE: should the name of the image be displayed on the image (default = TRUE)\cr
Only works with the 'raster' method. The label is also not displayed on saved images}

\item{label.txt}{text to be display on the image label. The same text will be displayed on all images (default = name of the image)}

\item{label.color}{color of the image label displayed on the image}

\item{label.x}{x position of the image name label}

\item{label.cex}{size of the text of the image name label (character expansion)}
}
\value{
the images are displayed or save but nothing is explicitly returned
}
\description{
Display or save an image
}
\keyword{display}
\keyword{image}
\keyword{label}
\keyword{save}
