% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCMapperFunctions.R
\name{getBorders}
\alias{getBorders}
\title{getBorders}
\usage{
getBorders(images)
}
\arguments{
\item{images}{list of images in data format}
}
\value{
list of images in data format where cell borders are indicated
}
\description{
Determine cell borders from the mask using a filter function
}
\keyword{border}
\keyword{cell}
\keyword{mask}
