% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCMapperFunctions.R
\name{getCellNumber}
\alias{getCellNumber}
\title{getCellNumber}
\usage{
getCellNumber(images)
}
\arguments{
\item{images}{list of images in data format}
}
\value{
list with number of cells per image
}
\description{
Returns a list of cell numbers (one list element per image)
}
\keyword{cell}
\keyword{number}
