% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCMapperFunctions.R
\name{getCells}
\alias{getCells}
\title{getCells}
\usage{
getCells(images, pop)
}
\arguments{
\item{images}{list of images in data format}

\item{pop}{ids of cells to be displayed}
}
\value{
list of images in data format where cells to be displayed are indicated
}
\description{
From a matrix of pixels corresponding to the mask image: pixels belonging to the cell population to be displayed = 1; other pixels = 0
}
\keyword{cells}
\keyword{image}
