% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCMapperFunctions.R
\name{getClusters}
\alias{getClusters}
\title{getClusters}
\usage{
getClusters(x, im.list, ids = FALSE)
}
\arguments{
\item{x}{data table with cell ids, image names, colors and clusters}

\item{im.list}{vector with names of images to be displayed}

\item{ids}{should ids (TRUE) or colors (FALSE) be returned}
}
\value{
a data table with either the cell ids or the colors to be displayed
}
\description{
Match each cell id to a color based on the cluster to which the cell belong
}
\keyword{cell}
\keyword{cluster}
\keyword{id}
