% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCMapperFunctions.R
\name{getCounts}
\alias{getCounts}
\title{getCounts}
\usage{
getCounts(x, ch, im.list)
}
\arguments{
\item{x}{data table with cell ids, channels and counts}

\item{ch}{channel to be displayed}

\item{im.list}{vector with names of images to be displayed}
}
\value{
data table with cell ids and respective counts for the channel to be displayed
}
\description{
Get the counts value for all cells in a given channel
}
\keyword{channel}
\keyword{counts}
