% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCMapperFunctions.R
\name{getIMCImages}
\alias{getIMCImages}
\title{getIMCImages}
\usage{
getIMCImages(image.metadata, ch1, ch2 = NULL, ch3 = NULL, path.col,
  file.col = NULL, slide.col, borders = FALSE)
}
\arguments{
\item{image.metadata}{path to the "Image.csv" file extracted from CellProfiler}

\item{ch1}{name of the 1st channel to display}

\item{ch2}{name of the 2nd channel to display}

\item{ch3}{name of the 3rd channel to display}

\item{path.col}{name of the column in the "Image.csv" file that contains the path to the masks}

\item{file.col}{name of the column in the "Image.csv" file that contains the names of the masks}

\item{slide.col}{name of the column in the "Image.csv" file that contains the image names}

\item{borders}{TRUE/FALSE should cell borders be displayed?}
}
\value{
a data table with links to the images and masks to be displayed
}
\description{
Recover the paths to the IMC tiff files from the "Image.csv" file and the image names provided by the user
}
\keyword{image}
\keyword{mask}
\keyword{path}
