% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCMapperFunctions.R
\name{getMasks}
\alias{getMasks}
\title{getMasks}
\usage{
getMasks(image.metadata, path.col, file.col, slide.col)
}
\arguments{
\item{image.metadata}{path to the "Image.csv" file extracted from CellProfiler}

\item{path.col}{name of the column in the "Image.csv" file that contains the path to the masks}

\item{file.col}{name of the column in the "Image.csv" file that contains the names of the masks}

\item{slide.col}{name of the column in the "Image.csv" file that contains the image names}
}
\value{
masks a data table with links to the masks
}
\description{
Recover the paths to the mask tiff files from the "Image.csv" file
}
\keyword{mask}
\keyword{path}
