% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCMapperFunctions.R
\name{makeLegend}
\alias{makeLegend}
\title{makeLegend}
\usage{
makeLegend(..., leg.names = c("1", "2", "3"), leg.fill = c("green", "red",
  "blue"), leg.border = c("black"), save = FALSE, path = getwd(),
  filename = "legend", fun = "", nb.items = 3)
}
\arguments{
\item{...}{additional parameters for the legend (see ?legend)}

\item{leg.names}{vector containing the names of the elements displayed in the legend}

\item{leg.fill}{vector containing the colors to be displayed in the legend}

\item{leg.border}{vector containing the color of the legend borders}

\item{save}{TRUE/FALSE should the legend be saved?}

\item{path}{path to the folder where legend should be saved}

\item{filename}{optional filename for the legend (note: should be modified and used as a suffix or prefix)}

\item{fun}{name of the parent function}

\item{nb.items}{number of items to be displayed in the legend}
}
\value{
the legend is displayed or saved, but nothing is explicitly returned
}
\description{
Create a legend as a separate image
}
\keyword{legend}
