% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCMapperFunctions.R
\name{MatchColorCluster}
\alias{MatchColorCluster}
\title{MatchColorCluster}
\usage{
MatchColorCluster(x, clusters, cluster.palette, return.x = TRUE)
}
\arguments{
\item{x}{data table with colors and clusters to display}

\item{clusters}{vector with cluster names}

\item{cluster.palette}{color palette to use, with one color per cluster}

\item{return.x}{should the data table (TRUE) or the lookup table (FALSE) be returned}
}
\value{
data table with clusters and matched color for each cluster
}
\description{
Match colors and clusters
}
\keyword{cluster}
\keyword{color}
