% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCMapperFunctions.R
\name{readImages}
\alias{readImages}
\title{readImages}
\usage{
readImages(pics, im.list, chan = "pic1")
}
\arguments{
\item{pics}{data table with links to the images and masks to display}

\item{im.list}{vector with names of images to be displayed}

\item{chan}{name of the column in pics containing the image to read}
}
\value{
list of images in data format
}
\description{
Read IMC images and store them in a list
}
\keyword{image}
\keyword{mask}
\keyword{read}
