% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMCMapperFunctions.R
\name{readMasks}
\alias{readMasks}
\title{readMasks}
\usage{
readMasks(masks, im.list)
}
\arguments{
\item{masks}{data table with links to the masks}

\item{im.list}{vector with names of images to be displayed}
}
\value{
list of images in data format
}
\description{
Read mask images (16-bit tiff) and store the image data in a list
}
\keyword{images}
\keyword{masks}
\keyword{read}
